fix: start background thread for identify rather than running it
The identify call is running the identify code synchronously instead of
starting the background thread.

This results in it blocking the calling JavaScript thread, essentially
making it a synchronous call even though it's returning an asynchronous
promise.

This fixes the typo so we start the thread which SHOULD make this
non-blocking for the calling react native app.