Add method which doesn't throw an exception when already initialised
isInitialized always throws an exception when already initialised. It seems like an anti-pattern to provide a boolean for that method which actually never can be true (as someone would expect). To prevent any breaking changes to others but also provide a more expected way of dealing with isInitialized, I would suggest to add a safe way which doesn't error if already initialised.