Add back the Android build step in CircleCI (#176)
Because of the Xcode version changes before, the CircleCI definition was
skipping the Android build. This added the build back and made sure the
step only ran on the latest RN versions so that the Gradle build would
work. I don't have an explanation of why CircleCI/RN 0.69 or RN 0.70
will run Gradle 7.5 but CircleCI/RN 0.71+ will run with Gradle 8.0.